Fix for bug 15202:  Application Qt-4.5.1 fails to configure when using open64 compiler
The problem is that guard variables for C++ local objects are not Initialized to 0 at the right program point.  When a constructor is created in a conditional expression, the flag is set to 1.  After the statement, a flag for an object is tested for 1 and destructor is called.  The bug was that in certain cases the intialization to 0 may not be executed resulting In subsequent code accessing unitialized value and calling destructor on an object that never got created. 

The fix is to mark the current top of stack as a place for guard variable Initialization and do the intialization in that block.  This problem can Manifest in C++ code with automatic destructors in if-then-else test expressions Or in ? Operator.

This fix will enhance the ability of the compiler to build complex C++ applications And Qt in particular.





git-svn-id: http://svn.open64.net/svnroot/open64@2700 ead328b1-e903-0410-a19d-eacee4b031f6