Bug 15476 -  Open64 bool return value ABI is not compatible with gcc
The CVTL operation was removed by Gen_stmt_wn() even if the CVTL has side
effect when the value is in the return register -- the upper part of the
return register may contain polluted bits. Added check to prevent the 
CVTL be removed if the value is in return register.




git-svn-id: http://svn.open64.net/svnroot/open64@2705 ead328b1-e903-0410-a19d-eacee4b031f6