Minor modifications to ConstraintGraphSolve to enable use for an IPA solution:   o Cycle detection is now dependent on the providing of a "local" constrain graph   o Cycle detction now ignores cross ConstraintGraph edges   o The postProcessPointsTo() moved out of ConstraintGraphSolve:solve()
ConstraintGraphNode::getCGNode() modified to prevent explosive generation of StInfo offsets.
We currently limit the number of offsets from a given ST to 32.  We may wish to tune this based
on datatype.  The StInfo class has been enhanced to track the max number of offsets on a
per-ST basis.
  
Provide skeleton driver for an IPA solution:  IPA_NystromAliasAnalyzer::solver()


git-svn-id: http://svn.open64.net/svnroot/open64@2919 ead328b1-e903-0410-a19d-eacee4b031f6