Enhance CFG transformation in WOPT to handle the CFG containing exception handling regions.
WOPT emitter generates WN based on CFG layout. Thus, in CFG transformation, 
the code should not be moved across the EH region boundaries. To achieve this, 

1) A BB_REGIONEXIT BB is added after an OPR_REGION when building CFG.  
After CFG is built, there is a pass of identifying the EH regions and mark 
every BB (rid_id) with its EH_region rid_id.

2) A new field layout_id is added in BB_NODE to represent where a BB should 
be placed. In clone_zone, when adding a new BB, the corresponding layout_id 
should be set right. 

3) When reconstructing CFG in CFG transformation, sorting BBs with their 
layout_id (if not 0) otherwise using their BB id to keep them in the right 
order with regards to EH-region requirements.

An assertion inside WOPT emitter is added to check the BB 's rid_id should 
be the right EH_region rid_id (to make sure the code layout is right).




git-svn-id: http://svn.open64.net/svnroot/open64@3138 ead328b1-e903-0410-a19d-eacee4b031f6