Fix bug 15666 - "numeric constant is out of range" for Fortran TINY.
On systems that use glibc-2.10.1, the use of the Fortran TINY intrinsic
with REAL(8) or REAL(4) types triggered a "numeric constant is out of
range" warning from the Fortran front-end.

The TINY values are specified in the front-end as string values, and the
warning is due to a difference in the way that the glibc routines strtof()
and strtod() detect underflow as compared to glibc 2.5.

To work around the issue, we replace the TINY constants with slightly
larger decimal string values so that the same binary value is returned by
each conversion function without setting errno to ERANGE.



git-svn-id: http://svn.open64.net/svnroot/open64@3139 ead328b1-e903-0410-a19d-eacee4b031f6