This change is to canonicalize loop-breaking condition.
The necessasity of this change can be explained by this example:

    for (int i = n; i < m; i++) {...}

IVR will introudce a PREG, say iv0, serving as primiary IV, making the original 
index <i> a secondary IV. In IVR, secondary IV will be substituted by an linear 
expr of primiary IV. Therefore, the loop-breaking condition becomes "iv0 + n < m".
This change is to canonicalize the expression into "iv0 < m-n". It would be 
easer to calcuate/estimate trip-count with the canonicaized form.

   LNO can live with the non-canonicaized form presumably thanks to 
class ACCESS_VECTOR which is used to depict LB/UB for individual loop.
Such class is certainly overkill for casual uses. 

   In this change, I only pick up a simple case to make me move on. I don't 
know for sure if it is always benefical to canonicalize loop-breaking 
condition, or to normalize loop, in IVR for any circumstances (For example, 
stride is not unit, or a symbolic var).



git-svn-id: http://svn.open64.net/svnroot/open64@3145 ead328b1-e903-0410-a19d-eacee4b031f6