Fix for 15718, 15719
The method Analyze_call() is called to decide if an edge can be inlined.  
It is also used for deleting an edge with the side effect of updating 
the call graph and the call counts associated with a caller.   So, care 
must be taken not to call this method more than once on an edge.

After the inliner checkin, this method was called multiple times on edges inside loops.   The fix is to ignore calls in loops after the these have been analyzed earlier.

This fixes the assert in Gamess and one of the output differences in 469.xercescbmk in spec rejects




git-svn-id: http://svn.open64.net/svnroot/open64@3179 ead328b1-e903-0410-a19d-eacee4b031f6