The fix in r975 was incorrect.  Inorder not call Analyze_call() twice, we were skipping calls in loops assuming that they were already processed by the new heuristic.  The problem is that the new heuristic may not see all calls in loops if allocated budget exceeded.  Yesterday's change will result in some edges not getting examined at all.  This has manifested as perf differences in gcc and libquantum and a failure to build mysql.
When we update the call graph after deleting an edge, we also mark the edge
that the caller has been updated.  We use two new methods

 IPA_EDGE::Has_Updated_Caller() : to query if caller of a deletable edge is
                                  updated

 IPA_EDGE::Set_Updated_Caller() : to mark a deletable edge that the caller
                                  is updated.

In Analyze_call(), when we encounter an edge that is deletable but hasn't
updated the callgraph, we update the call graph and mark the edge.  Later
when we encounter the same edge, we ignore it becuase update is already done.

This changes allows one to call Analyze_call() multiple times without causing
inconsistency in call graph.





git-svn-id: http://svn.open64.net/svnroot/open64@3180 ead328b1-e903-0410-a19d-eacee4b031f6