Fix for 15614:  GemsFDTD fails with prefetch=0.
With the factorization change we ended up unrolling more 
than before after invariant factorization.  The fix is to unroll 
after factorization only if the current loop contains 
a loop with  delay_unroll flag.  This is similar to the earlier behavior
before factorization change.  Should help with compile time as well.




git-svn-id: http://svn.open64.net/svnroot/open64@3203 ead328b1-e903-0410-a19d-eacee4b031f6