Fix bug 15358: wrong register used for inline asm instruction.
This bug was found while trying to compile ffmpeg with Open64 at -O2.

For an input to inline assembly bound to the expression
"((unsigned char)(-s))", the compiler used %ecx (32 bits) instead of
%cl (8 bits) in the generated assembly.

The root cause was that the size was not directly stored, only implied
by a "U4CVTL 8" operation used to load the asm input and this operation
could be moved by the optimizer.

The fix, which had already been made in the PSC fork of the source, is
to introduce a temporary with the correct size in wgen.



git-svn-id: http://svn.open64.net/svnroot/open64@3206 ead328b1-e903-0410-a19d-eacee4b031f6