Fix for defect 15415: gamess asserts in r461.
We were always rebuilding dependence graph after factorization
because of a potential full unroll.   In some cases, 
the program is so large that we assert that some of 
constants used in dependence graph builder are no longer
sufficient.

The current fix is to rebuild only when it is necessary.
If we fully unrolled after factorization, we record this 
fact and rebuild only in that case.

In future we may have to revisit the dependence graph builder
to gracefully handle larger programs. 




git-svn-id: http://svn.open64.net/svnroot/open64@3208 ead328b1-e903-0410-a19d-eacee4b031f6