Fix bug 15358 (again): wrong register used for inline asm instruction.
A previous bug fix worked for -O2.  A similar problem was found while
trying to compile ffmpeg at -O3: an input to inline assembly with the
type of "unsigned char" was being promoted to register %ecx instead of
the expected %cl.

A side effect of scalar_rename is that the type of the renamed variable
is promoted to 32 bits.  The fix is to exclude inputs to inline assembly
from renaming when the type would be promoted.



git-svn-id: http://svn.open64.net/svnroot/open64@3217 ead328b1-e903-0410-a19d-eacee4b031f6