Fix bug 15820: out of memory for application boost-1-38-0/wave
The problem is running out of memory in cfg transformation when
adding paths for conditional constant.

In function find_conditional_const, it could possibly process the
same coderep (and add its paths) for the number of BB in cfg.
The fix is to add a cr_visited bvector to record CRs that are visited.


git-svn-id: http://svn.open64.net/svnroot/open64@3234 ead328b1-e903-0410-a19d-eacee4b031f6