Avoid doing forward substitution when uses are inline asm inputs.
The exact type of each asm input must be preserved in order to generate
the correct instructions.  Forward substitution may have the side-effect
of promoting the type and therefore needs be skipped in this case.

An earlier attempt to fix the issue was made in trunk commit r2318.
The problem with the previous fix is that it tried to skip the FS in
FS_Substitute(), which is too late.  By then FS_Worthwhile() has
returned TRUE indicating that is possible to do forward substitution
and delete the original store, which FS_Substitute() does
unconditionally.

In the "draw_edges_mmx" test case extracted from ffmpeg (see bug 539),
the combination of the deleted store and skipped substitution resulted
in an asm input that was loaded from uninitialized memory.

Instead, FS_Worthwhile() should return FALSE if any of the uses are
directly under an OPR_ASM_INPUT node.

Approved by: Sun Chan


git-svn-id: http://svn.open64.net/svnroot/open64@3322 ead328b1-e903-0410-a19d-eacee4b031f6