Avoid bash-specific redirection operators in make rules.
For the top-level clean action, clean both 'lib' and 'lib2' targets if
they are in use.  Do not clean GNU3 components if not building them.

For libhugetlbfs, ensure that make does not try to build the 'version'
file in the source tree using an implicit rule for creating executables.

Approved by: Suneel Jain


git-svn-id: http://svn.open64.net/svnroot/open64@3343 ead328b1-e903-0410-a19d-eacee4b031f6