This is a fix for a recent regression which is seen with virtual destructors.
The issue is that we are not removing dead code (and unreachable code)
in WOPT in case the if-test contains EXTRACT_BITS. So we are introducing
a simp_node function that bypasses EXTRACT_BITS when it is
extracting the last bit of a Boolean constant. This plugs in to the
simp_node framework and further simplifications are also possible,
in the future. 

The assembly with and without the patch are
given below for the following test case. 
Please look at the destructor functions:
_ZN9TestClassD1Ev, _ZN9TestClassD2Ev for the dead code.

Test case:
class TestClass
{
        virtual ~TestClass();
};


TestClass::~TestClass()
{
}

Asm with fix: 
_ZN9TestClassD1Ev:	# 0x10
    # .frame	%rsp, 24, %rsp
.LBB1__ZN9TestClassD1Ev:
    movq $(_ZTV9TestClass+16),%rax	# [0] _ZTV9TestClass+16
    movq %rax,0(%rdi)             	# [1] id:7
    .loc	1	10	0
 #   9  {
 #  10  }
    ret                           	# [1] 
.LDWend__ZN9TestClassD1Ev:

Asm without the fix:
tClassD1Ev:	# 0x1c
    # .frame	%rsp, 24, %rsp
    .loc	1	8	0
.LBB1__ZN9TestClassD1Ev:
    addq $-24,%rsp                	# [0] 
    xorl %esi,%esi                	# [0] 
    movq $(_ZTV9TestClass+16),%rax	# [1] _ZTV9TestClass+16
    .loc	1	10	0
    testl $1,%esi                 	# [1] 
    .loc	1	8	0
    movq %rax,0(%rdi)             	# [2] id:7
    .loc	1	10	0
    jne .LBB2__ZN9TestClassD1Ev   	# [2] 
.Lt_1_1282:
    addq $24,%rsp                 	# [0] 
    ret                           	# [0] 
    .p2align 5,,31
.LBB2__ZN9TestClassD1Ev:
    call _ZdlPv                   	# [0] _ZdlPv
.LBB4__ZN9TestClassD1Ev:
    jmp .Lt_1_1282                	# [0] 
.LDWend__ZN9TestClassD1Ev:




git-svn-id: http://svn.open64.net/svnroot/open64@3415 ead328b1-e903-0410-a19d-eacee4b031f6