Use pattern rules for make recipes that produce multiple targets.
This is the documented way to ensure that GNU make handles this kind of
target correctly for a parallel make.

The previous rule for the generated files targ_si.[ch] was incorrect
and could cause a build failure when using make with the '-j' option,
because make would try to build targ_si.h and targ_si.c by invoking the
generator two times in parallel.

Approved by: Jian-Xin Lai


git-svn-id: http://svn.open64.net/svnroot/open64@3486 ead328b1-e903-0410-a19d-eacee4b031f6