Fix for 727.
There have been several changes to routine WGEN_Expand_Expr (wgen_expr.cxx)
having to do with gspin trees below GS_INIT_EXPR  nodes (see the comments in
WGEN_Expand_Expr in the GS_MODIFY_EXPR case and the comments for "Bug 11253"
and bug 15210").  These changes (and others) modify the gspin input tree to
make it easier to translate to WHIRL.  This problem occurs in the case where
the gspin tree is:

Original gspin              Modified to
-------------------------------------------------
GS_INIT_EXPR
  GS_COMPONENT_REF (A)
  GS_TARGET_EXPR
    GS_VAR_DECL             -> GS_COMPONENT_REF (A)
    GS_COND_EXPR
      GS_NE_EXPR
      GS_TARGET_EXPR
        GS_VAR_DECL         -> GS_COMPONENT_REF (A)
        <operand 1>
      GS_TARGET_EXP
        GS_VAR_DECL         -> GS_COMPONENT_REF (A)
         <operand 1>

The problem occurs because, in the context of a GS_COND_EXPR, the GS_TARGET_EXP
operand 1 is not expanded so no symbol table entry is built for the expression
value. This results in the dereference of a NULL symbol table entry pointer.
In the original problem report, operand 1 is a GS_CONSTRUCTOR.  Subsequent
testing revealed that the problem occurs for cases when operand 1 something
other than a GS_CONSTRUCTOR.

This change addresses the problem by expanding operand 1.



git-svn-id: http://svn.open64.net/svnroot/open64@3497 ead328b1-e903-0410-a19d-eacee4b031f6