Fix bug 735: C++ exception handling introduces false catch-all regions.
The root cause of the problem is that the two cases of catch-all (no
type filter) and cleanup (landing pad but no handler, and therefore no
type) are not distinguished in the compiler's internal representation
of EH info.  A "hack" used to tell the difference in the CG appears to
have incorrect assummptions as it is easy to generate a test case where
a cleanup is marked as a catch-all in the output range table, and will
thus fail at runtime when the C++ runtime treats the cleanup landing
pad as if it were a handler.

This fix uses the formerly unused INITV "value" of INITVKIND_ONE to
definitively represent the catch-all case.

- In wgen, rename "append_catch_all" to "append_cleanup".  This routine
  appends a potential cleanup, not a catch-all marker.  Use the type
  filter value of INITVKIND_ONE for real catch-all regions.
- In the cg, the detection of catch-all is simplified and no guesswork
  is needed.
- Adjust code that writes or copies the internal EH info to expect
  INITVKIND_ONE as a possible type filter value.

Note: further changes to the TARG_IA64-specific implementation of
Create_INITO_For_Range_Table are required to complete the fix for
that target.

Approved by: Jian-Xin Lai


git-svn-id: http://svn.open64.net/svnroot/open64@3518 ead328b1-e903-0410-a19d-eacee4b031f6