Improve optimization for C++ programs.
1. Multi-target icall promotion

The implementation has two parts.  The first part is in IPA summary
phase, which generates multiple dummy call sites for the indirect
call target based on the feedback info: if the number of times
of the callsite is called exceeds the IPA_Icall_Min_Freq, and the
rate of one callee among the total calls of the callsite exceeds
IPA_Icall_Target_Min_Rate, then a dummy callsite is added for the callee.
Up to ICALL_MAX_PROMOTE_PER_CALLSITE dummy callsites could be added.

The second part is in IPA transformation phase, which generates code
based on the number of dummy callsites coming with the indirect call,
and the wn_icall is transformed to

      if (*fptr == callee1)
         callee1(...);
      else {
         if (*fptr == callee2)
            callee2(...);
         else {
            ...
            fptr(...);
         }
      }

When a new call is added, the feedback info is updated with corresponding
call frequency, and the callsite map is updated with newly created call
node.

2. Devirtualization with overridden map

The current devirtualization mechanism doesn't build virtual function
overridden relationship and use it to direct devirtualization,
which causes a lot of devirtualization opportunities to be missed.
By building the overridden map, we can safely devirtualize the virtual
call unconditionally in whole program mode when there is only one
candidate available in the overridden map for the call.

3. Fully_Unroll_Short_Loops bug fix

The first two changes exposed a case where Fully_Unroll_Short_Loops
was called recursively on a deleted WHIRL node.  The fix is to add a
parameter that indicates whether the caller is iterating over the WHIRL
nodes sequentially so that Fully_Unroll_Short_Loops can avoid visiting
"WN_next(wn)".

Approved by: Min Zhao and Sun Chan


git-svn-id: http://svn.open64.net/svnroot/open64@3559 ead328b1-e903-0410-a19d-eacee4b031f6