This change is a kludge in an attempt to reduce the overhead of auto-vectorization. The overhead we are trying to reduce is the cost of computing lower bound (LB) and upper bound (UB) of vectorized loop and remainder loop.
Suppose the original loop is:
   for (i = 0; i < n; i++)  and the unrolling factor (i.e. vector length) is f.

After SIMD, the vectorized loop and the remainder loop were the following:
   vectorized loop: for (i = 0; i <= (n/f)*f - 1; i+=f) {}
   remainder loop : for (i =(n/f)/f; i < n; i++) {}

The cost of computing LB/UB is high if :
  - the loop being vectorized is in a deep nest, and 
  - the trip-count is small and it is not known at compile time

With this change, the vectorized loop and remainder loop are now:
   vectorized loop: for (i = 0; i <= n-f; i+=f) {}
   remainder loop : for (i'= i; i' < n; i'++) {}

The change to lnopt_hoistif.cxx is to prevent the LB of the remainder loop 
from going back to what it was. The "hoist if" phase blindly changes the 
code in order to not let index variable live out of loop. The change to 
this file is to prevent loops from being modified if they are not amenable 
to hoist-if optimization.

Also, this change keeps index variable private to the loop being vectorized
if the loop is flagged Do_Loop_Is_MP().

Code review by Mei Ye.


git-svn-id: http://svn.open64.net/svnroot/open64@3598 ead328b1-e903-0410-a19d-eacee4b031f6