Remove some inline declarations in omp_rtl.h
If using gcc > 4.3, it used C99 inline rules. This caused some extern inline
functions declared in omp_rtl.h to be not defined.  But using -fgnu89-inline
breaks the build for older versions of gcc. Instead, I remove the inline
keyword for declaration of the following functions in omp_rtl.h:

__ompc_init_vp
__ompc_task_delete
__ompc_task_get
__ompc_task_switch

git-svn-id: http://svn.open64.net/svnroot/open64@3633 ead328b1-e903-0410-a19d-eacee4b031f6