This patch introduces a framework for vectorization of IF-statements of the form:   - if (x != 0) { single_istore_statement } else {empty_body}   - if (x != 0) { if (y !=0) {single_istore_statement} else {empty_body}} else {empty_body} This optimization is turned on by default and can be controlled by -LNO:simd_vect_if={on/off}.
Sample program:
long array[2000000];
void sample() {
  long i;
  for(i = 0; i < 2000000; i++) {
    if (array[i])
	array[i] ^= i;
  }
  return;
}

Vectorization of such an if-statement is done by first vectorizing its subexpressions.
The result of the vectorized if-condition is computed by 'pcmpeqq' (V16I8V16I8EQ) and
is used to select between the result of executing the statement in the if-body or 
leaving the array element unchanged. This selection is performed by the 
'pblendvb' (V16I1V16I1SELECT) operation. Both of these are SSE4.1 instructions.
This patch extends the supported types of
 - OPR_SELECT to include V16I1 (to support generation of pblendvb operation).
 - OPR_EQ to include V16I8 (to support generation of pcmpeqq operation).

This patch also recognizes and handles vectorization of invariants rooted at OPR_ADD, OPR_SUB and OPR_MPY.

C.R. by Fred Chow, Mei Ye and Jian-Xin Lai.



git-svn-id: http://svn.open64.net/svnroot/open64@3681 ead328b1-e903-0410-a19d-eacee4b031f6