Fix bug #876.
This patch is related to bug 827 and rev 3739. For a simple case:
extern void undef();
static void foo() {
  undef();        // <-- A
}
static void bar() {
  int n;
  switch(n) {
    foo();        // <-- B
  }
}
int global() {
  bar();
}

The callsite B is dead code and eliminated by the GCC FE at O2. With the
patch to bug 827, function foo is still marked as referenced. Then
function foo() is emitted to the BE and generate the call to "undef".
There will be an error in the final linking phase: undefined reference
to `undef()'. In the previous patch rev 3739, when we generate the CALL node to a
function, both the name and decl of the function are marked as
referenced. That causes this regression. Since the logic to detect if a
static function is used but not defined only check the name of the
symbol, the new patch only set the name of the function referenced.

Code Review by Sun Chan.



git-svn-id: http://svn.open64.net/svnroot/open64@3747 ead328b1-e903-0410-a19d-eacee4b031f6