Adds support for taskwait and collapse in Fortran FE
Note that it doesn't check if collapse count exceeds number of perfectly
nested loops in the front-end, but back-end was modified so that in this
case it will only collapse up to the available loops.

This also fixes a bug in lastprivate clause when collapse clause is also
used. It fixes the check for whether the executing thread executed the last
iteration when collapse clause is present.

Contributor: Deepak

git-svn-id: http://svn.open64.net/svnroot/open64@3807 ead328b1-e903-0410-a19d-eacee4b031f6