[CAF] fix to allow use of uhcaf in makefiles
The uhcaf script will run the make command to determine the communication libs
from GASNet and ARMCI makefiles. If uhcaf is invoked from within another
makefile, the output of the nested makefile will include a printing of working
directory. To avoid this, we use --no-print-directory command in the script.

Also, there was a bug during coarray prelowering phase (undefined variable
ty1) that was fixed.

Contributor: Deepak

git-svn-id: http://svn.open64.net/svnroot/open64@3885 ead328b1-e903-0410-a19d-eacee4b031f6