[CAF] Fixes file name DST info generated in wgen
If compiling a source code not in the current directory with -g, it couldn't
be stepped through in a debugger. It turns out that the DWARF symbol table
(DST) information for the file name was not appending the relative path. This
prevents the relative path from being stripped by wgen before writing out the
DST information. This fix enables us to build all the runtime libraries (e.g.
libfortran) with debug info.

Also:
- adds --keep option for uhcaf script
- increases patch level, now at 3.0.3

Contributor: Deepak

git-svn-id: http://svn.open64.net/svnroot/open64@3889 ead328b1-e903-0410-a19d-eacee4b031f6