This is a fix for a bug found by adding new fields to the OP structure which caused a segmentation fault during the code generation phase when building SPEC dealII.
The Base() member function of a POINTS_TO structure only
returns a valid ST pointer when the base kind is fixed (for
debugging purposes, during alias information construction
pointers to different structures are placed in _base when
the base kind is dynamic).

For consistency sake, in this source file I changed the
other check to determine whether the base kind is fixed to
use the Based_is_fixed() member function.

CR: Jian-Xin Lai.


git-svn-id: http://svn.open64.net/svnroot/open64@3898 ead328b1-e903-0410-a19d-eacee4b031f6