fix for bug952.
In the expansion of OPR_STID for X8664 -m32 target,  
when the kid is 64bit result with paired TN but the stid_type
is 4byte size or less, we delete the useless higher 32bit tn 
by unpair the result. This will help pass the pair check in the 
following expansion routines.

Code Review: Lai Jian-Xin.


git-svn-id: http://svn.open64.net/svnroot/open64@3906 ead328b1-e903-0410-a19d-eacee4b031f6