Fix bug 779: build in source directory does not work.
In the osprey-gcc-4.2.0 build always use a known subdirectory for
$(host_subdir) instead of making the choice depending on where configure
was run, so that we always know where to find the build artifacts.

Also in the osprey-gcc 4.2.0 build use a path relative to $(build_objdir)
to find libgspin42.a.

In the build of the cygnus/ld subdirectory, use the "new-ld" target
instead of "all".  The "all" target causes unwanted modifications to
checked-in files when the build is being done in the source directory.
For Open64, we only need the new-ld (ipa_link) executable.

Approved by: Sun Chan and Mike Murphy


git-svn-id: http://svn.open64.net/svnroot/open64@3925 ead328b1-e903-0410-a19d-eacee4b031f6