[CAF] Supports save coarrays + misc. fixes
The main change in this commit is added support for save coarrays. In order to
efficiently support remotely accessible save coarrays, they must be allocated
on the managed symmetric heap as allocatable coarrays are. This requires the
compiler to gather all save coarrays declared throughout the program and
allocate them on the heap at the start of the program. OpenUH now does this.

This commit also fixes a number of bugs:

* Fix for created LCB assignment statement during the coarray prelowering
  phase

* Fix for uhcaf return status and command line option parsing

* Fixes for handling OPR_CVT WHIRL nodes during coarray lowering phase.

* Fixes bug when writing a scalar value to a remote array section.

* Corrects the data type for a remote access within structured control flow
  statements

* Fixes dope vector temp variable created for calling a procedure with an
  assumed shape coarray dummy argument. Specifically, the temp needed to be
  declared with codimensions matching the dummy argument.

Also, increasing patch level to 3.0.7

Contributor: Deepak

git-svn-id: http://svn.open64.net/svnroot/open64@3939 ead328b1-e903-0410-a19d-eacee4b031f6