Fix omp_lib.h/omp_lib and add configure option for default CAFRT heap size.
The omp_lib.h file can now be included in F90 source files. Also, the external
names for OpenMP library routines in OMP_LIB.mod now match the names defined
in libopenmp (removed the second underscore).

This commit also adds a new configure option for setting the default PGAS heap
size in the CAF runtime, --with-default-caf-heap-size.  It is possible to set
the default CAF heap size with the new configure option.  If not specified,
the default default value is 30M.

Contributor: Deepak

git-svn-id: http://svn.open64.net/svnroot/open64@3975 ead328b1-e903-0410-a19d-eacee4b031f6