[CAF] Adds preliminary support for F2008 locks
This adds preliminary support for F2008 locks. The LOCK and UNLOCK statements
may be used to acquire/release a lock variable on any image. The runtime
algorithm is adapted from the MCS lock algorithm.

This also adds the necessary runtime support for some atomic operations --
swap, fetch-and-store, fetch-and-add, and compare-and-swap (GASNet). Atomics
are implemented with GASNet using active messages. ARMCI provides swap and
fetch-and-add atomics.

Contributor: Deepak, Tony

git-svn-id: http://svn.open64.net/svnroot/open64@3977 ead328b1-e903-0410-a19d-eacee4b031f6