Adds more checks for explicit Fortran interface + lock fixes
Give error if no explicit interface for subroutines with allocatable arg.  The
compiler would not give a warning or error when calling a subroutine with an
allocatable dummy argument but no explicit interface. Instead, it would
generate bad code (passing an uninitialized address instead of the dope
vector). This adds an error check for subroutines without an explicit
interface when the dummy has allocatable, optional, or value attribute.

We also add a necessary error check for allocatable coarray dummy with
intent(out).

This commit also fixes a lock/unlock data race.  The runtime implementation
would deadlock in some scenarios before. This is because when checking for an
update to the local lock request descriptor, it was waiting for both the image
and ofst fields to be updated. We expand the lock_request_t type to include a
sentinel field done, which will be set to 1 once image and ofst have been
updated.

Finally, to allow for asynchronous servicing the AM requests, this
incorporates the OpenSHMEM progress thread implementation into the CAF
runtime. It creates an additional thread which will poll the network and
handle new AM requests periodically.

Contributors: Deepak, Tony

git-svn-id: http://svn.open64.net/svnroot/open64@3978 ead328b1-e903-0410-a19d-eacee4b031f6