[CAF] Use PSHM directly instead of AM for atomic operations with GASNet
Compute node information is filled into the nodeinfo_table during comm_init.
This allows us to determine if a target process resides on the same compute
node, in which case we can directly perform an atomic update on its memory
rather than using an active message. This benefits performance and also allows
for 1-sided progress within a compute node for atomic operations, without
requiring a spawned progress thread.

Before I was using a mutex for all atomic updates to a given process. Now, we
use the GNU builtins for atomics instead. Atomic variables may be a 1-byte,
2-byte, 4-byte, or 8-byte object.

Contributor: Deepak

git-svn-id: http://svn.open64.net/svnroot/open64@3989 ead328b1-e903-0410-a19d-eacee4b031f6