[CAF] Fix enabling of progress thread
If using the infiniband conduits (ibv or vapi), we utilize GASNET's internal
progress thread support instead of spawning our own. Note that this will only
work for AMs sent over the network. So, on the infiniband conduits an AM sent
within a node will not progress until the target's client code has serviced
the request. In the future, we may add a routine that can be invoked in the
CAF code (or inserted by the compiler) to ensure progress.

We also fix the trace for service/progress thread which is used for
non-infiniband GASNet conduits.

Contributor: Deepak

git-svn-id: http://svn.open64.net/svnroot/open64@3990 ead328b1-e903-0410-a19d-eacee4b031f6