[CAF] fix reservation of symmetric heap + misc script improvements
This commit fixes reservation of the symmetric heap space, particularly when
attempting to reserve more than 2 GB of heap space.  There were a number fo
issues that needed to be fixed:

* In a recent change, when I read in the environment variable for
  UHCAF_IMAGE_HEAP_SIZE if it exceeded 4 GB (32-bit size) the code would read
  it as 0. Needed to replace a call to atoi with atoll.

* In the general case, I was setting GASNET_MAX_SEGSIZE to a value less than
  UHCAF_IMAGE_HEAP_SIZE when UHCAF_IMAGE_HEAP_SIZE > 2000000000.

* And even if GASNET_MAX_SEGSIZE  = UHCAF_IMAGE_HEAP_SIZE, its not enough
  because GASNet appears to reserve 2 MB of the segment for something.

Now, it reserves enough space so that GASNET_MAX_SEGSIZE will
always be a page-size multiple and at least 2 MB greater than what was
requested.  During runtime, instead of giving an error if the
UHCAF_IMAGE_HEAP_SIZE is still greater than the maximum segment size, it will
display a warning and use the lesser value.

The cafrun script has been improved. Options provided by the underlying
launcher (e.g. mpirun) can be specified after the "--" at the end of the
command line. These options will be joined with the existing options specified
in LAUNCHER_OPTS.

More:

* --help menu only displays the help options. To see the environment
    variables, use --help-env.
* adds a --version option.
* --show will print the executable name and the communication layer and
  network it will run on.
* -np is accepted for specifying number of images  (same as -n and
  --num-images=)

Further, improvements were made in uhcaf compile script and cafrun launch
script to automatically identify which MPI libraries need to be linked in and
also which launcher to use if unspecified by the user. Its tested with Open
MPI and MVAPICH. When compiling, instead of using the variable LIBMPI to
specify the libraries to link in, the variable MPI_LD_LIBS is used.

Our current implementation imposes a limit on number of images (1024*1024) and
heap size per image (64GB). This commit adds a checks for these values during
initialization.

MPI_INCLUDE needs to be set to build the runtime against the ARMCI libraries.
The configure script will now set the MPI include path if it finds mpicc in
the user's path.  When compiling a CAF program with ARMCI layer, MPI_LIB needs
to be set. The compile script will similarly set MPI_LIB if it finds mpicc in
the path.

Contributor: Deepak

git-svn-id: http://svn.open64.net/svnroot/open64@3996 ead328b1-e903-0410-a19d-eacee4b031f6