[CAF] fix hanging during startup for CAF+OMP programs
The OMP runtime library uses a static constructor to initialize the initial
team of threads. The presense of a team of threads for each image appears to
disrupt the GASNet and ARMCI initialization. So, now it does the following:

* if the libopenmp static constructor detects the presence of the weak symbol
  __caf_init, it will not initialize the thread team.

* if __caf_init, in libcaf, detects the presence of the weak symbol
  __ompc_init_rtl, it will call it before exiting.

The end result is that if both libcaf and libopenmp are linked in, the
creation of the thread team will be postponed until after GASNet or ARMCI
initialization has completed.

This commit also fixes a bug in the OMP runtime library relating to
non-deferred tasks (if clause evaluates to 0).

Contributor: Deepak

git-svn-id: http://svn.open64.net/svnroot/open64@4005 ead328b1-e903-0410-a19d-eacee4b031f6