[CAF] support pointer components for coarrays, and some FE checks
The use of pointer components in coarrays of derived types should now be
supported when using the GASNet communication layer. When using ARMCI, pointer
components may be used, but can not be assigned to a target with the "=>"
operation (i.e. they may only be allocated/deallocated).

Accessing a remote object with the target attribute is possible using GASNet
active messages. The support added for this derives from the corresponding
implementation for accessing remote global variables in the UH OpenSHMEM
library.

This commit also adds some FE checks that were missing before.

* Adds a check for non-contiguous coarray actual arguments that are
  associated with non-assumed-shape coarray dummy arguments. This is not
  allowed becase it would require an implicit synchronization. An appropriate
  error message is now given in this case.

* Check that allocatable coarray components are deferred-shape, which the
  compiler was overlooking before.

* If not compiling with GASNet, the FE will give an error if attempting to do
  a pointer assignment to a pointer component of a coarray.

There was also a fix in the uhcaf script for compiling with the gasnet-udp
conduit.

Contributors: Deepak, Tony

git-svn-id: http://svn.open64.net/svnroot/open64@4007 ead328b1-e903-0410-a19d-eacee4b031f6