[CAF] Fix for sync images + CAFRT debug build
* The _SYNC_IMAGES routine in CAF runtime was modifying the image_list array
argument.  This argument should be read only.

* Need to pass -DCAFRT_DEBUG if BUILD_OPTIMIZE = DEBUG.  Also, only print file
and line number on warning and errors if its a debug build of the runtime.

* Adjusting how GASNET_MAX_SEGSIZE gets set. It should be set to at least 2MB
greater than the requested heap size. Still not clear what decides the
difference between GASNET_MAX_SEGSIZE and the maximum local segment size, but
2MB is a lower bound. Using 2MB * 8 in addition to the requested heapsize
(adjusted to 4K page boundary) for now, which hopefully should be enough.

Contributor: Deepak

git-svn-id: http://svn.open64.net/svnroot/open64@4025 ead328b1-e903-0410-a19d-eacee4b031f6