[CAF] support for stat= and errmsg= on synchronization statements
This adds compiler support for the stat= and errmsg= specifers on sync
statements (sync all, sync images, and sync memory) and lock and unlock
statements. Usage is:

- sync all [([stat=..|errmsg=..])]
- sync images (image-list, [stat=..|errmsg=..])
- sync memory [([stat=..|errmsg=..])]
- lock [lock-var, ([acquired_lock=..|stat=..|errmsg=..])]
- unlock [lock-var, ([stat=..|errmsg=..])]

Currently, only blank error messages are returned but his may be changed in
the future (the error messages are implementation-defined according to the
standard).

The STAT_STOPPED_IMAGE status code is returned for sync all and sync images if
trying to synchronize with a statement that has begun termination. The current
implementation will probably not scale well, and we are considering more
efficient approaches for communicating stopped images to synchronization
partners.

The lock statement will return STAT_LOCKED if trying to lock a variable that
is already held The unlock statement will return STAT_UNLOCKED if trying to
unlock a variable that is already unlocked. It will return
STAT_LOCKED_OTHER_IMAGE if trying to unlock a variable that is locked by a
different image.

Contributor: Deepak

git-svn-id: http://svn.open64.net/svnroot/open64@4028 ead328b1-e903-0410-a19d-eacee4b031f6