[CAF] requires ARMCI installation instead of ARMCI src directory
Instead of the user giving the configure option --with-armci-src-root=, they
will give --with-armci-root=<armci-installation> for using CAF on top of
ARMCI. Additional linker arguments required for the particular network that
will be used may also be specified with --with-armci-network-libs=<...>. For
example, for an Infiniband network, the user might add this to the configure
line:

--with-armci-root=/path/to/armci/install --with-armci-network-libs="-libverbs"

It would be better if there was a way to figure out which libraries need to be
linked in from the installation directory, but the ARMCI installation doesn't
currently provide that information.

Also updating patch level to 3.0.20.

Contributor: Deepak

git-svn-id: http://svn.open64.net/svnroot/open64@4029 ead328b1-e903-0410-a19d-eacee4b031f6