VHO fix for privatizing temporaries in MP region
When VHO generates temporary variables during vho_region_lower, it needs to
ensure that they are privatized (i.e. a PRAGMA_LOCAL for the temporary is
added to the regions pragma block). This functionality was added before by
Jiangzhou He in the Open64 openmp3.0 branch, but it wasn't merged in correctly
into OpenUH.

Contributors: Jiangzhou, Deepak

git-svn-id: http://svn.open64.net/svnroot/open64@4032 ead328b1-e903-0410-a19d-eacee4b031f6