[CAF] Fortran front-end fix for pointer assignment
This corrects a mistake introduced in 3.0.20 which resulted in the front-end
crashing when processing a pointer assignment statement. The uhcaf script will
now set OPENUH_COMM_LAYER, and the front-end will read this to decide whether
to support pointer assignment to components of derived-type coarrays (only
allowed when linking against GASNet layer).

Updating patch level to 3.0.23

git-svn-id: http://svn.open64.net/svnroot/open64@4034 ead328b1-e903-0410-a19d-eacee4b031f6